Registry redux
No more RegistryHandler. Now there's a Registry interface which is identified by a URL and knows how to get the MODULE.bazel file, how to get the fetch info, and how to actually fetch a module. Different implementations of Registry deal with different schemes (so http,https,file are "registry indexes", fake://foo is a fake registry used by tests, and later we could add git+https:// as a single module registry), so you can get a Registry object just from a URL with a scheme. And registry.GetModuleBazel knows how to deal with the default list of registries and an override, and returns the Registry object that was actually used.