Archive fetcher mega commit
This commit implements the Archive fetcher (formerly known as the HTTP fetcher; renamed as it takes care of local file:// archives too). It makes use of the central HTTP download cache and the shared repos directory, and is cognizant of vendoring mode.

Hopefully the API of the fetch module is clear enough now that I can start plumbing through the flags and actually take it for a test drive.