Archive fetcher: require fingerprint file for shared repo dirs, too
Even though the name of a shared repo dir is the same as the fingerprint, having the fingerprint file is still a useful signal since we don't write the fingerprint until everything else has succeeded. Thus checking the fingerprint file allows us to avoid using a half-finished download from earlier.